pyproject.toml
[tool.poetry]
name = "beryl"
version = "0.1.0"
description = "custom bot for silicas server, coded by honchokomodo, and improved by No767"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
py-cord = {version = "2.0.0rc1", extras = ["voice"]}
orjson = "^3.6.8"
aiodns = "^3.0.0"
brotlipy = "^0.7.0"
cchardet = "^2.1.7"
python-dotenv = "^0.20.0"
uvloop = "^0.16.0"
requests = "^2.27.1"
pysimdjson = "^4.0.3"
disquest-utils = "^1.0.4"
ffmpeg-python = "^0.2.0"
beanie = "1.11.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
